requirement.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.4
PyJWT==2.8.0
passlib[argon2]==1.7.4